move obstype change to fix tests
